Премодерация - запрос на изменение
git-svn-id: svn://dev.cliffa.net/ausweis@536 6fdb692e-5d52-e011-adc2-000e0c3ec3fd

+++ lib/DB/Preedit.pm
@@ -92,7 +92,6 @@ sub add {
                 op      => 'E',
                 recid   => $args{recid},
                 modered => 0,
-                '`cnt`' => 0,
             } ,
             { 
                 join => 'field',
@@ -101,6 +100,7 @@ sub add {
             },
         );
         foreach my $p (@p) {
+            next if $p->{cnt} || !defined($p->{cnt});
             $self->schema->model('Preedit')->delete({ id => $p->{id} })
                 || return;
         }
